Amelioration du ProcessPaiement
